requirements.txt
planet-dsl==1.0.9
fastapi
uvicorn[standard]
pandas
